+++ packages/api/models/schemas.js
@@ -216,8 +216,8 @@ module.exports.collection = {
         description: 'chunk size of the S3 multipart uploads for the collection',
         type: 'number',
       },
-      metadataFileExtension: {
-        title: 'Optional, the file extension for tasks such as move-granule to look for the granule metadata information',
+      granuleMetadataFileExtension: {
+        title: 'Optional file extension for tasks such as move-granule to look for the granule metadata information, e.g. .cmr.json, .iso.xml',
         type: 'string',
       },
     },